Add indexing for configurable products
Configurable product are constructed by having a super product
and adding associated products to it with differing super product
attribute configurations.

The implementation finds the associated products and the configurable
attributes and adds their values as multivalue fields to the super
product.

Associated products are not being indexed them selves.