[BUGFIX] Remove query terms from showing mini form
Problem is that the mini form is cached and thus may show old query
terms. Since the mini form is shown on every page you likely do not
want to make it uncached either. The current query term should rather
be shown in a separate form that belongs to the results page.